+++ src/tabview/HISTORY.md
@@ -4,7 +4,10 @@ TabView Change History
 @VERSION@
 ------
 
-* No changes.
+* Fixed missing ARIA role in the `tablist` of Tabview.
+  ([#1035][]: @blzaugg)
+
+[#1035]: https://github.com/yui/yui3/issues/1035
 
 3.11.0
 ------