+++ src/tabview/HISTORY.md
@@ -1,6 +1,14 @@
 TabView Change History
 ======================
 
+3.12.0
+------
+
+* Fixed missing ARIA role in the `tablist` of Tabview.
+  ([#1035][]: @blzaugg)
+
+[#1035]: https://github.com/yui/yui3/issues/1035
+
 3.11.0
 ------